Revise and test functions in io_data.py. Post issues...

+++ mhealthx/io_data.py
@@ -11,19 +11,19 @@ Copyright 2015,  Sage Bionetworks (http://sagebase.org), Apache v2.0 License
 
 
 def read_synapse_table_files(synapse_table_id,
-                             column_name='', select_rows=[], output_path='.',
+                             column_names=[], max_row=None, output_path='.',
                              username='', password=''):
     """
-    Read data from a Synapse table. If column_name specified, download files.
+    Read data from a Synapse table. If column_names specified, download files.
 
     Parameters
     ----------
     synapse_table_id : string
         Synapse ID for table
-    column_name : string
-        column header for fileIDs in Synapse table (if wish to download files)
-    select_rows : list
-        row indices, if retrieving column_name files (empty means all rows)
+    column_names : list of strings
+        column headers for columns with fileIDs (if wish to download files)
+    max_row : int
+        number of rows to retrieve files from (None means all rows)
     output_path : string
         output path to store column_name files
     username : string
@@ -33,24 +33,25 @@ def read_synapse_table_files(synapse_table_id,
 
     Returns
     -------
-    dataframe : Pandas DataFrame
+    table_data : Pandas DataFrame
         Synapse table contents
-    files : list of strings
-        files from Synapse table (full paths to downloaded files)
+    files : list of lists of strings
+        files from Synapse table column(s) (full paths to downloaded files)
 
     Examples
     --------
     >>> from mhealthx.io_data import read_synapse_table_files
-    >>> synapse_table_id = 'syn4590865'
-    >>> column_name = 'audio_audio.m4a'
-    >>> select_rows = range(3)  # [] for all rows
+    >>> synapse_table_id = 'syn4590865' #'syn4907789'
+    >>> column_names = ['audio_audio.m4a', 'audio_countdown.m4a']
+    >>> max_row = 3  # None for all rows
     >>> output_path = '.'
     >>> username = ''
     >>> password = ''
-    >>> dataframe, files = read_synapse_table_files(synapse_table_id, column_name, select_rows, output_path, username, password)
+    >>> table_data, files = read_synapse_table_files(synapse_table_id, column_names, max_row, output_path, username, password)
 
     """
     import synapseclient
+    import numpy as np
 
     syn = synapseclient.Synapse()
 
@@ -60,38 +61,57 @@ def read_synapse_table_files(synapse_table_id,
     else:
         syn.login()
 
-    # Download Synapse table schema and dataframe:
+    # Download Synapse table schema and table_data:
     schema = syn.get(synapse_table_id)
-    results = syn.tableQuery("select * from {0}".format(synapse_table_id))
-    dataframe = results.asDataFrame()
+    query = "select * from {0}".format(synapse_table_id)
+    results = syn.tableQuery(query)
+    table_data = results.asDataFrame()
     files = []
 
-    # Download Synapse files for fileIDs in specified table column:
-    if column_name:
-        if not select_rows:
-            select_rows = range(dataframe.shape[0])
-        for rowID in select_rows:
-            fileinfo = syn.downloadTableFile(schema,
-                                             rowId=rowID,
-                                             versionNumber=0,
-                                             column=column_name,
-                                             downloadLocation=output_path)
-            if fileinfo:
-                files.append(fileinfo['path'])
-            else:
-                files.append('')
+    # Download Synapse files for fileIDs in specified table columns:
+    if column_names:
 
-    return dataframe, files
+        # Set the number of rows to loop through:
+        if not max_row:
+            max_row = table_data.shape[0]
+
+        # Loop through specified columns:
+        for column_name in column_names:
+            column_data = table_data[column_name]
+
+            # Loop through specified number of rows:
+            files_per_column = []
+            for irow in range(max_row):
+
+                # If there is no file in that row, save an empty string:
+                if np.isnan(column_data[irow]):
+                    files_per_column.append('')
+
+                # Download the file:
+                else:
+                    fileinfo = syn.downloadTableFile(schema,
+                                                 rowId=irow,
+                                                 versionNumber=0,
+                                                 column=column_name,
+                                                 downloadLocation=output_path)
+                    if fileinfo:
+                        files_per_column.append(fileinfo['path'])
+                    else:
+                        files_per_column.append('')
+
+            files.append(files_per_column)
+
+    return table_data, files
 
 
-def write_synapse_table(dataframe, synapse_project_id, schema_name='',
+def write_synapse_table(table_data, synapse_project_id, schema_name='',
                         username='', password=''):
     """
     Write data to a Synapse table.
 
     Parameters
     ----------
-    dataframe : Pandas DataFrame
+    table_data : Pandas DataFrame
         Synapse table contents
     synapse_project_id : string
         Synapse ID for project within which table is to be written
@@ -104,14 +124,17 @@ def write_synapse_table(dataframe, synapse_project_id, schema_name='',
 
     Examples
     --------
-    >>> from mhealthx.io_data import read_synapse_table, write_synapse_table
+    >>> from mhealthx.io_data import read_synapse_table_files, write_synapse_table
     >>> input_synapse_table_id = 'syn4590865'
     >>> synapse_project_id = 'syn4899451'
-    >>> dataframe = read_synapse_table(input_synapse_table_id, username, password)
-    >>> schema_name = 'Contents of ' + input_synapse_table_id
+    >>> column_names = []
+    >>> max_row = None
+    >>> output_path = '.'
     >>> username = ''
     >>> password = ''
-    >>> write_synapse_table(dataframe, synapse_project_id, schema_name, username, password)
+    >>> table_data, files = read_synapse_table_files(input_synapse_table_id, column_names, max_row, output_path, username, password)
+    >>> schema_name = 'Contents of ' + input_synapse_table_id
+    >>> write_synapse_table(table_data, synapse_project_id, schema_name, username, password)
 
     """
     import synapseclient
@@ -125,16 +148,17 @@ def write_synapse_table(dataframe, synapse_project_id, schema_name='',
     else:
         syn.login()
 
-    dataframe.index = range(dataframe.shape[0])
+    table_data.index = range(table_data.shape[0])
 
-    schema = Schema(name=schema_name, columns=as_table_columns(dataframe),
+    schema = Schema(name=schema_name, columns=as_table_columns(table_data),
                     parent=synapse_project_id, includeRowIdAndRowVersion=False)
 
-    syn.store(Table(schema, dataframe))
+    syn.store(Table(schema, table_data))
 
 
-def copy_synapse_table(synapse_table_id, synapse_project_id, schema_name='',
-                       dataframe=None, remove_columns=[], username='', password=''):
+def copy_synapse_table(synapse_table_id, synapse_project_id,
+                       schema_name='', table_data=None, remove_columns=[],
+                       username='', password=''):
     """
     Copy Synapse table contents to another Synapse project.
 
@@ -143,13 +167,13 @@ def copy_synapse_table(synapse_table_id, synapse_project_id, schema_name='',
     synapse_table_id : string
         Synapse ID for table to copy
     synapse_project_id : string
-        Synapse ID for project within which table is to be written
+        copy table to project with this Synapse ID
     schema_name : string
         schema name of table
-    dataframe : Pandas DataFrame
+    table_data : Pandas DataFrame
         Synapse table contents (optional)
     remove_columns : list of strings
-        column headers for columns not to be copied
+        column headers for columns to be removed
     username : string
         Synapse username (only needed once on a given machine)
     password : string
@@ -157,7 +181,7 @@ def copy_synapse_table(synapse_table_id, synapse_project_id, schema_name='',
 
     Returns
     -------
-    dataframe : Pandas DataFrame
+    table_data : Pandas DataFrame
         Synapse table contents
     schema_name : string
         schema name of table
@@ -169,11 +193,12 @@ def copy_synapse_table(synapse_table_id, synapse_project_id, schema_name='',
     >>> from mhealthx.io_data import copy_synapse_table
     >>> synapse_table_id = 'syn4590865'
     >>> synapse_project_id = 'syn4899451'
-    >>> schema_name = 'mPower phonation feature file table'
+    >>> schema_name = 'Copy of ' + synapse_table_id
+    >>> table_data = None
     >>> remove_columns = ['audio_audio.m4a', 'audio_countdown.m4a']
     >>> username = ''
     >>> password = ''
-    >>> dataframe, schema_name, synapse_project_id = copy_synapse_table(synapse_table_id, synapse_project_id, schema_name, remove_columns, username, password)
+    >>> table_data, schema_name, synapse_project_id = copy_synapse_table(synapse_table_id, synapse_project_id, schema_name, table_data, remove_columns, username, password)
 
     """
     import synapseclient
@@ -189,26 +214,28 @@ def copy_synapse_table(synapse_table_id, synapse_project_id, schema_name='',
         syn.login()
 
     # Download Synapse table as a dataframe:
-    if dataframe.empty:
+    if table_data is None:
         results = syn.tableQuery("select * from {0}".format(synapse_table_id))
-        dataframe = results.asDataFrame()
+        table_data = results.asDataFrame()
 
     # Remove specified columns:
-    for remove_column in remove_columns:
-        del dataframe[remove_column]
+    if remove_columns:
+        for remove_column in remove_columns:
+            del table_data[remove_column]
 
     # Upload to Synapse table:
-    dataframe.index = range(dataframe.shape[0])
-    schema = Schema(name=schema_name, columns=as_table_columns(dataframe),
+    table_data.index = range(table_data.shape[0])
+    schema = Schema(name=schema_name, columns=as_table_columns(table_data),
                     parent=synapse_project_id,
                     includeRowIdAndRowVersion=False)
-    table = syn.store(Table(schema, dataframe))
+    table = syn.store(Table(schema, table_data))
 
-    return dataframe, schema_name, synapse_project_id
+    return table_data, schema_name, synapse_project_id
 
 
 def upload_files_handles_to_synapse(input_files, synapse_project_id,
-                                    schema_name='', username='', password=''):
+                                    schema_name='', column_name='fileID',
+                                    username='', password=''):
     """
     Upload files and file handle IDs to Synapse.
 
@@ -220,6 +247,8 @@ def upload_files_handles_to_synapse(input_files, synapse_project_id,
         Synapse ID for project within which table is to be written
     schema_name : string
         schema name of table
+    column_name : string
+        header for column of fileIDs
     username : string
         Synapse username (only needed once on a given machine)
     password : string
@@ -228,15 +257,18 @@ def upload_files_handles_to_synapse(input_files, synapse_project_id,
     Examples
     --------
     >>> from mhealthx.io_data import upload_files_handles_to_synapse
-    >>> input_files = ['/Users/arno/Local/wav/test1.wav', '/Users/arno/Local/wav/test2.wav']
+    >>> input_files = ['/Users/arno/Local/wav/test1.wav']
     >>> synapse_project_id = 'syn4899451'
     >>> schema_name = 'Test to store files and file handle IDs'
+    >>> column_name = 'fileID1'
     >>> username = ''
     >>> password = ''
-    >>> upload_files_handles_to_synapse(input_files, synapse_project_id, schema_name, username, password)
+    >>> upload_files_handles_to_synapse(input_files, synapse_project_id, schema_name, column_name, username, password)
+    >>> #column_name = 'fileID2'
+    >>> #input_files = ['/Users/arno/Local/wav/test2.wav']
+    >>> #upload_files_handles_to_synapse(input_files, synapse_project_id, schema_name, column_name, username, password)
 
     """
-    import os
     import synapseclient
     from synapseclient import Schema
     from synapseclient.table import Column, RowSet, Row
@@ -255,60 +287,42 @@ def upload_files_handles_to_synapse(input_files, synapse_project_id,
         file_handle = syn._chunkedUploadFile(input_file)
         files_handles.append([file_handle['id']])
 
-    # Upload files and file handle IDs:
-    cols = [Column(name='fileID', columnType='FILEHANDLEID')]
-    schema = syn.store(Schema(name=schema_name, columns=cols, parent=synapse_project_id))
-    syn.store(RowSet(columns=cols, schema=schema,
+    # New column headers:
+    new_column_header = Column(name=column_name, columnType='FILEHANDLEID')
+
+    # See if Synapse table exists:
+    # tex = list(syn.chunkedQuery("select id from Table where parentId=='{0}'"
+    #                             " and name=='{1}'".format(synapse_project_id,
+    #                                                       schema_name)))
+    # If Synapse table does not exist, create table schema:
+    # if not tex:
+
+    # Create table schema:
+    schema = syn.store(Schema(name=schema_name,
+                              columns=[new_column_header],
+                              parent=synapse_project_id))
+
+    # Upload files and file handle IDs with new schema:
+    syn.store(RowSet(columns=[new_column_header], schema=schema,
                      rows=[Row(r) for r in files_handles]))
 
-
-def append_file_names(input_files, file_append):
-    """
-    Copy each file with an append to its file name.
-
-    Parameters
-    ----------
-    input_files : list of strings
-        full path to the input files
-    file_append : string
-        append to each file name
-
-    Returns
-    -------
-    output_files : list of strings
-        each string is the full path to a renamed file
-
-    Examples
-    --------
-    >>> from mhealthx.io_data import append_file_names
-    >>> input_files = ['/Users/arno/mhealthx_working/mHealthX/phonation_files/test.tmp']
-    >>> file_append = '.m4a'
-    >>> output_files = append_file_names(input_files, file_append)
-
-    """
-    import os
-    from shutil import copyfile
-
-    output_files = []
-
-    # Loop through input files:
-    for input_file in input_files:
-        if not os.path.exists(input_file):
-            raise(IOError(input_file + " not found"))
-        else:
-            # Don't copy file if file already has correct append:
-            if input_file.endswith(file_append):
-                output_files.append(input_file)
-            # Copy to a new file with appended file name:
-            else:
-                output_file = input_file + file_append
-                copyfile(input_file, output_file)
-                if not os.path.exists(output_file):
-                    raise(IOError(output_file + " not found"))
-                else:
-                    output_files.append(output_file)
-
-    return output_files
+    # If Synapse table exists, get table ID and update table schema:
+    # else:
+    #
+    #     # Download Synapse table schema and table_data:
+    #     schema = syn.get(synapse_table_id)
+    #     results = syn.tableQuery("select * from {0}".format(synapse_table_id))
+    #     table_data = results.asDataFrame()
+    #
+    #     synapse_table_id = tex[0]['Table.id']
+    #     schema = syn.get(synapse_table_id)
+    #     new_column = syn.store(new_column_header)
+    #     schema.addColumn(new_column)
+    #     schema = syn.store(schema)
+    #
+    #     # Upload files and file handle IDs with new schema:
+    #     syn.store(RowSet(columns=[new_column_header], schema=schema,
+    #                      rows=[Row(r) for r in files_handles]))
 
 
 def convert_audio_files(input_files, file_append, command='ffmpeg',
@@ -391,29 +405,19 @@ if __name__ == '__main__':
     target_synapse_project_id = 'syn4899451'
     username = ''
     password = ''
-    column_name = 'audio_audio.m4a'
-    select_rows = range(3) #[]  # Test with first 3 rows: range(3)
+    column_name = 'audio_audio.wav'
+    max_row = 3 # Test with first 3 rows or None for all rows
     output_path = '.'
     ffmpeg = '/home/arno/software/audio/ffmpeg/ffmpeg'
     schema_name = 'mPower phonation wav files and file handle IDs'
     #schema_name = 'mPower countdown wav files and file handle IDs'
 
     # Download files:
-    dataframe, files = read_synapse_table_files(synapse_table_id, column_name,
-                                                select_rows, output_path,
-                                                username, password)
-
-    # Rename files with .m4a append (copy):
-    file_append1 = '.m4a'
-    renamed_files = append_file_names(files, file_append1)
-
-    # Convert from m4a to wav:
-    file_append2 = '.wav'
-    input_args = '-i'
-    output_args = '-ac 2'
-    output_files = convert_audio_files(renamed_files, file_append2, ffmpeg,
-                                       input_args, output_args)
+    table_data, files = read_synapse_table_files(synapse_table_id,
+                                                 column_name, max_row,
+                                                 output_path,
+                                                 username, password)
 
     # Upload wav files and file handle IDs:
-    upload_files_handles_to_synapse(output_files, target_synapse_project_id,
+    upload_files_handles_to_synapse(files, target_synapse_project_id,
                                     schema_name, username, password)
